+++ test/test_surface_flux.jl
@@ -1,7 +1,3 @@
-using Oceananigans.Biogeochemistry: AbstractContinuousFormBiogeochemistry
-import Oceananigans.Biogeochemistry: required_biogeochemical_tracers
-
-using Walrus.RadiativeTransfer: HomogeneousBodyHeating, PARModelHeating
 using Walrus.SurfaceFluxModel: OceanAtmosphereBoundaryConditions, 
                                PrescribedAtmosphericState, 
                                EmpiricalDownwellingLongwave,
@@ -9,29 +5,8 @@ using Walrus.SurfaceFluxModel: OceanAtmosphereBoundaryConditions,
                                heat_exchange_coefficient,
                                SimilarityTheoryInterface
 
-struct JustPhytoplankton <: AbstractContinuousFormBiogeochemistry end
-required_biogeochemical_tracers(::JustPhytoplankton) = (:P, )
-
-@testset "Surface heat exchange" begin
-    grid = RectilinearGrid(arch; extent = (1, 1, 10), size = (1, 1, 10))
-
-    #####
-    ##### Test body heating
-    #####
-
-    body_heating = HomogeneousBodyHeating(; surface_flux = (x, y, t) -> 1,
-                                            water_attenuation_coefficient = 1.,
-                                            water_density = 1.,
-                                            water_heat_capacity = 1.)
-
-    model = NonhydrostaticModel(; grid, forcing = (; T = Forcing(body_heating, discrete_form=true)), tracers = :T)
-
-    time_step!(model, 10)
-
-    # 1 W/m² for 10s -> 10J, top meter absorbs 10J * (exp(0) - exp(- 1)) ≈ 6.3212 J
-    # Density and specific head are 1 so the water should have increased in temp by ~6.3212 K
-    @test Array(interior(model.tracers.T, 1, 1, 10))[1] ≈ 10 * (1 - exp(-1))
-
+@testset "Surface flux (heat and momentum)" begin
+   grid = RectilinearGrid(arch; extent = (1, 1, 10), size = (1, 1, 10))
     #####
     ##### Test radiative heating/cooling (no sensible or latent flux)
     #####
@@ -186,54 +161,43 @@ required_biogeochemical_tracers(::JustPhytoplankton) = (:P, )
     @test Array(interior(model.tracers.T, 1, 1, 10))[1] > -1
 
     #####
-    ##### OceanBioME light attenuation body heating
+    ##### Momentum transfer
     #####
-
     grid = RectilinearGrid(arch; size = (2, 2, 2), extent = (2, 2, 2))
 
-    light_attenuation_model = TwoBandPhotosyntheticallyActiveRadiation(; grid)
-    body_heating = PARModelHeating(; light_attenuation_model)
+    atmosphere = PrescribedAtmosphericState(wind_speed = 0.0, wind_direction = 90.0, temperature = 0.0)
 
-    biogeochemistry = Biogeochemistry(JustPhytoplankton(); light_attenuation = light_attenuation_model)
+    U, V, Q = OceanAtmosphereBoundaryConditions(grid, atmosphere)
 
-    model = NonhydrostaticModel(; grid, biogeochemistry, timestepper = :QuasiAdamsBashforth2, forcing = (; T = Forcing(body_heating, discrete_form=true)), tracers = :T)
+    model = NonhydrostaticModel(; grid, 
+                                  boundary_conditions = (u = FieldBoundaryConditions(top = U),
+                                                         v = FieldBoundaryConditions(top = V)),
+                                  tracers = :T)
 
-    Pᵢ(x, y, z) = 2.5 + z
+    for n=1:10
+        time_step!(model, 1)
+    end
 
-    set!(model, P = Pᵢ)
+    @test all(Array(interior(model.velocities.u)) .≈ 0) & all(Array(interior(model.velocities.v)) .≈ 0) # no wind no stress
 
-    kʳ = light_attenuation_model.water_red_attenuation
-    kᵇ = light_attenuation_model.water_blue_attenuation
-    χʳ = light_attenuation_model.chlorophyll_red_attenuation
-    χᵇ = light_attenuation_model.chlorophyll_blue_attenuation
-    eʳ = light_attenuation_model.chlorophyll_red_exponent
-    eᵇ = light_attenuation_model.chlorophyll_blue_exponent
-    r = light_attenuation_model.pigment_ratio
-    Rᶜₚ = light_attenuation_model.phytoplankton_chlorophyll_ratio
+    atmosphere = PrescribedAtmosphericState(wind_speed = 1.0, wind_direction = 90.0, temperature = 0.0)
 
-    zc = znodes(grid, Center(), Center(), Center())
+    U, V, Q = OceanAtmosphereBoundaryConditions(grid, atmosphere)
 
-    Δz = zspacings(grid, Center())
+    model = NonhydrostaticModel(; grid, 
+                                  boundary_conditions = (u = FieldBoundaryConditions(top = U),
+                                                         v = FieldBoundaryConditions(top = V)),
+                                  tracers = :T)
 
-    Chlʳ = [(Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eʳ, (Pᵢ(0, 0, zc[1]) * Rᶜₚ / r) ^ eʳ]
-    Chlᵇ = [(Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eᵇ, (Pᵢ(0, 0, zc[1]) * Rᶜₚ / r) ^ eᵇ]
+    time_step!(model, 1)
+    
+    for n=1:10000
+        time_step!(model, 1)
+    end
 
-    ∫Chlʳ = [(Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eʳ * Δz[1, 1, 1]/2]
-    ∫Chlᵇ = [(Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eᵇ * Δz[1, 1, 1]/2]
-
-    push!(∫Chlʳ, ∫Chlʳ[1] + (Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eʳ * Δz[1, 1, 1]/2 + (Pᵢ(0, 0, zc[1]) * Rᶜₚ / r) ^ eʳ * Δz[1, 1, 1]/2)
-    push!(∫Chlᵇ, ∫Chlᵇ[1] + (Pᵢ(0, 0, zc[2]) * Rᶜₚ / r) ^ eᵇ * Δz[1, 1, 1]/2 + (Pᵢ(0, 0, zc[1]) * Rᶜₚ / r) ^ eᵇ * Δz[1, 1, 1]/2)
-
-    expected_PAR = 100.0 .* [exp(zc[2] * kʳ - ∫Chlʳ[1] * χʳ) + exp(zc[2] * kᵇ - ∫Chlᵇ[1] * χᵇ),
-                             exp(zc[1] * kʳ - ∫Chlʳ[2] * χʳ) + exp(zc[1] * kᵇ - ∫Chlᵇ[2] * χᵇ)] ./ 2
-
-    analytical_body_heating = 100.0 .* [(kʳ + Chlʳ[1] * χʳ) * exp(zc[2] * kʳ - ∫Chlʳ[1] * χʳ) + (kᵇ + Chlᵇ[1] * χᵇ) * exp(zc[2] * kᵇ - ∫Chlᵇ[1] * χᵇ),
-                                        (kʳ + Chlʳ[2] * χʳ) * exp(zc[1] * kʳ - ∫Chlʳ[2] * χʳ) + (kᵇ + Chlᵇ[2] * χᵇ) * exp(zc[1] * kᵇ - ∫Chlᵇ[2] * χᵇ)] ./ 2
-
-    analytical_body_heating ./= (body_heating.water_density * body_heating.water_heat_capacity)
-
-    # high tollerance, I think error is from course grid (hopefully)
-    @test CUDA.@allowscalar all(isapprox.(analytical_body_heating, [body_heating(1, 1, k, grid, model.clock, fields(model)) for k = 2:-1:1], atol = 1.5e-6))
+    @test -1 <= Array(interior(model.velocities.u, 1, 1, 2))[1] < 0
+    @test Array(interior(model.velocities.u, 1, 1, 1))[1] ≈ 0
+    @test all(Array(interior(model.velocities.v)) .≈ 0)
 end
 #= TODO:
 @testset "Similarity theory interface coefficients" begin
